clippy: fix clippy warnings.
Fix clippy warning:

error: casting to the same type is unnecessary (usize -> usize)
--> image/src/cmdline.rs:73:28
|
73 | let str = args[idx as usize].clone();
| ^^^^^^^^^^^^ help: try: idx
|
= help: for further information visit https://rust-lang.github.io/rust-clippy/master/index.html#unnecessary_cast
= note: -D clippy::unnecessary-cast implied by -D warnings

+++ image/src/cmdline.rs
@@ -70,7 +70,7 @@ impl ArgsParse {
         let mut pre_opt = (0, "".to_string());
 
         for idx in 0..len {
-            let str = args[idx as usize].clone();
+            let str = args[idx].clone();
             if str.starts_with('-') && str.len() > 1 {
                 if !pre_opt.1.is_empty() {
                     bail!("missing argument for option '{}'", pre_opt.1);
